feat: Add support for captcha solving
Zeit introduced a FriendlyCaptcha that needs to be solved in order to log in.
We solve this captcha using ati-captcha.com and save the received session cookies
on disk. If we find the session file again, we will reuse those cookies in order
to avoid solving too many captchas.

It appears that session cookies are valid for a year.